requirements.txt
Flask
Flask-SQLAlchemy
fpdf2
Flask-Login
Werkzeug
pytz
gunicorn
psycopg2-binary